requirements.txt
arabic-reshaper==2.0.15
asgiref==3.6.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-cleanup==8.0.0
django-cors-headers==3.14.0
django-crispy-forms==1.7.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
future==1.0.0
grpcio==1.65.5
grpcio-status==1.65.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.8
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==23.0
pillow==10.2.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydub==0.25.1
PyJWT==2.6.0
pyparsing==3.1.2
pypng==0.20220715.0
pyproj==3.6.1
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
reportlab==4.1.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
SpeechRecognition==3.8.1
sqlparse==0.4.3
typing_extensions==4.10.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
